requirements.txt
django
flask
numpy
joblib
scikit-learn
django-storages[azure]
azure-storage-blob
celery
django
numpy
opencv-python
tensorflow
pillow
djangorestframework
streamlit
python-dotenv
groq